requirements.txt
APScheduler==3.11.0
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.0
cssselect2==0.8.0
Flask==3.1.1
fonttools==4.58.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzlocal==5.3.1
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
gevent==23.9.1
python-dotenv==1.0.0

